+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from "./bolt11";
 export * from "./lnurl";
 export * from "./podcasting2";
-export * from "./l402";
+export * from "./402/l402";
 export * from "./fiat";
